refactor: change validation
