+++ src/main/java/br/com/icorrespondencia/api/dto/TownhouseDTO.java
@@ -1,7 +1,15 @@
 package br.com.icorrespondencia.api.dto;
 
+import java.util.List;
+
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotEmpty;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonView;
+
+import br.com.icorrespondencia.api.domain.Block;
+import br.com.icorrespondencia.api.dto.validation.View;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -22,6 +30,7 @@ import lombok.experimental.SuperBuilder;
 @EqualsAndHashCode(callSuper = false)
 @NoArgsConstructor(access = AccessLevel.PACKAGE)
 @AllArgsConstructor(access = AccessLevel.PACKAGE)
+@JsonView(View.Public.class)
 public class TownhouseDTO extends PersonDTO {
 
     private static final String ENTER_A_VALID_MAIL_VALIDATION_MESSAGE = "Enter a valid email";
@@ -33,4 +42,8 @@ public class TownhouseDTO extends PersonDTO {
 
     @Email(message = ENTER_A_VALID_MAIL_VALIDATION_MESSAGE, regexp = "^[A-z0-9\\.]+@[A-z0-9]+\\.[A-z]{1,}\\.?[A-z]+$")
     private String email;
+
+    private String phone;
+
+    private List<Block> blocks;
 }
