+++ src/main/java/br/com/icorrespondencia/api/dto/UserDTO.java
@@ -3,8 +3,6 @@ package br.com.icorrespondencia.api.dto;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotEmpty;
 
-import br.com.icorrespondencia.api.exception.ValidationGroups;
-
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -28,12 +26,10 @@ public class UserDTO extends PersonDTO {
 
     private static final String ENTER_A_VALID_MAIL_VALIDATION_MESSAGE = "Enter a valid email";
 
-    @NotEmpty(message = FIELD_MUST_BE_FILLED_VALIDATION_MESSAGE,
-            groups = { ValidationGroups.Post.class, ValidationGroups.Put.class })
+    @NotEmpty(message = FIELD_MUST_BE_FILLED_VALIDATION_MESSAGE)
     private String username;
 
-    @NotEmpty(message = FIELD_MUST_BE_FILLED_VALIDATION_MESSAGE,
-            groups = { ValidationGroups.Post.class, ValidationGroups.Put.class })
+    @NotEmpty(message = FIELD_MUST_BE_FILLED_VALIDATION_MESSAGE)
     private String password;
 
     @Email(message = ENTER_A_VALID_MAIL_VALIDATION_MESSAGE, regexp = "^[A-z0-9\\.]+@[A-z0-9]+\\.[A-z]{1,}\\.?[A-z]+$")
