+++ src/main/java/br/com/icorrespondencia/api/dto/validation/View.java
@@ -1,0 +1,10 @@
+package br.com.icorrespondencia.api.dto.validation;
+
+public interface View {
+
+    interface Public {
+    }
+
+    interface Internal extends Public {
+    }
+}
